Inline more methods in SelectorFilter https://bugs.webkit.org/show_bug.cgi?id=273558 rdar://127375992
Reviewed by NOBODY (OOPS!).

Instrumenting page load time with Dtrace shows a large number of calls into SelectorFilter methods
during initial page load and rendering. Specifically, during about 1s of loading, there are
approximately 40k calls to both SelectorFilter::chooseSelectorHashesForFilter() and
SelectorFilter::collectHashes(). This makes these two methods the #22 and #23 most frequent
methods called in WebCore during a representative page load.

* Source/WebCore/WebCore.xcodeproj/project.pbxproj:
* Source/WebCore/css/SelectorFilter.cpp:
(WebCore::SelectorFilter::collectSelectorHashes): Deleted.
(WebCore::SelectorFilter::chooseSelectorHashesForFilter): Deleted.
(WebCore::SelectorFilter::collectHashes): Deleted.
* Source/WebCore/css/SelectorFilter.h:
(WebCore::SelectorFilter::fastRejectSelector const): Deleted.
* Source/WebCore/css/SelectorFilterInlines.h: Added.
(WebCore::SelectorFilter::fastRejectSelector const):
(WebCore::SelectorFilter::collectSelectorHashes):
(WebCore::SelectorFilter::chooseSelectorHashesForFilter):
(WebCore::SelectorFilter::collectHashes):
* Source/WebCore/css/parser/CSSTokenizerInputStream.h:
(WebCore::CSSTokenizerInputStream::peek const):
* Source/WebCore/style/ElementRuleCollector.cpp:
* Source/WebCore/style/HasSelectorFilter.cpp:
* Source/WebCore/style/RuleData.cpp:
