+++ Source/WebCore/style/ElementRuleCollector.cpp
@@ -44,6 +44,7 @@
 #include "SVGElement.h"
 #include "SelectorCheckerTestFunctions.h"
 #include "SelectorCompiler.h"
+#include "SelectorFilterInlines.h"
 #include "SelectorMatchingState.h"
 #include "ShadowRoot.h"
 #include "StyleProperties.h"
